docs: moved wallet signer docs snippets
